LayoutTests/imported/w3c/web-platform-tests/infrastructure/expected-fail/unhandled-rejection-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: error outside any setup or test

Harness Error (FAIL), message = Unhandled rejection: error outside any setup or test
